+++ src/Types/LogLevel.ts
@@ -1,0 +1,9 @@
+/* 
+    Enum for log level
+    Default is 'info'
+*/
+export enum LogLevel {
+    Log_Info = 'info',
+    Log_Warning = 'warning',
+    Log_Error = 'error'
+};