+++ src/Types/ProjectTemplateType.ts
@@ -1,7 +1,0 @@
-
-
-export enum ProjectTemplateName {
-    ConsoleApplication = 'ConsoleApplication',
-    DynamicLibrary = 'DynamicLibrary',
-    SharedLibrary = 'SharedLibrary'
-};